pyproject.toml
[project]
name = "DocuMind"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiomysql>=0.3.2",
    "argon2-cffi>=25.1.0",
    "chromadb>=1.5.7",
    "cryptography>=46.0.7",
    "fastapi>=0.135.3",
    "langchain>=1.2.15",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.1",
    "langchain-text-splitters>=1.1.1",
    "openai>=2.31.0",
    "passlib>=1.7.4",
    "pydantic[email]>=2.12.5",
    "pypdf>=6.10.0",
    "python-dotenv>=1.2.2",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.26",
    "sentence-transformers>=5.4.0",
    "sqlmodel>=0.0.38",
    "uvicorn>=0.44.0",
]
